getCameras() átkerült a RecognizerHandlerbe a loadchain miatt
