+++ App/Models/MouseProcessor.py
@@ -151,16 +151,16 @@ class MouseProcessor:
         #Mutatóujj
         if angle1 > 45:
             if now - self.last_click_time > timedelta(seconds=0.4):
-                self.mouse.press(Button.right if self.invert else Button.left)
-                self.mouse.release(Button.right if self.invert else Button.left)
+                self.mouse.press(Button.left if self.invert else Button.right)
+                self.mouse.release(Button.left if self.invert else Button.right)
                 self.last_click_time = now
             return
       
         #Hüvelykujj
         if angle2 > 70:
             if now - self.last_click_time > timedelta(seconds=0.4):
-                self.mouse.press(Button.left if self.invert else Button.right)
-                self.mouse.release(Button.left if self.invert else Button.right)
+                self.mouse.press(Button.right if self.invert else Button.left)
+                self.mouse.release(Button.right if self.invert else Button.left)
                 self.last_click_time = now
             return
